no error when no pulse cookie

+++ entrypoint.sh
@@ -11,7 +11,9 @@ mkdir -p /mpd /mpd/db /mpd/pid /mpd/state /mpd/sticker /mpd/log /mpd/music /mpd/
 chown "${USER_NAME}:${USER_NAME}" /mpd /mpd/db /mpd/pid /mpd/state /mpd/sticker /mpd/log /mpd/music /mpd/playlists
 
 gosu "${USER_NAME}" mkdir -p "/home/${USER_NAME}/.config/pulse"
-ln -s /pulseaudio/cookie "/home/${USER_NAME}/.config/pulse/cookie"
+if [ -f /pulseaudio/cookie ]; then
+  ln -s /pulseaudio/cookie "/home/${USER_NAME}/.config/pulse/cookie"
+fi
 
 gosu "${USER_NAME}" envsubst '${MPD_PASSWORD}' < /mpd/mpd.conf.template > /mpd/mpd.conf
